objectids "cache" option
added option to configuration called "queryAllFeaturesMaxAgeInMilliseconds". This is the number of milliseconds to cache the list of all objectIds. This is to save time because on large datasets where many of the features do not have attachments, this query may happen multiple times each time a new location is loaded, slowing down the application. This variable is set in configuration.json and should generally be set to {boolean} false unless you are having loading speed issues when switching between images.